% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLinearModel.R
\name{plotLinearModel}
\alias{plotLinearModel}
\title{Plots data with linear model}
\usage{
plotLinearModel(data, x, y)
}
\arguments{
\item{data}{Dataset in form of tibble that is being plotted (Must have at least three valid row)}

\item{x}{Column in data that is the x axis of plot (format of string)}

\item{y}{Column in data that is the y axis of plot (format of string)}
}
\value{
ggplot2 scatter plot with linear model and data points with linear
function as the title and adjusted r-squared and F-statistic as subtitle
}
\description{
Plot quantitative data from column x and y in data in a scatter plot along with
a linear model function of the two column.
}
\examples{
test_tibble <- tibble::tribble(
~x, ~y, ~z,
#--|--|----
1, 2, 3,
2, 3, 4,
3, 5, 2,
4, 3, 2
)
plotLinearModel(test_tibble, 'x', 'y')

large_data <- tibble::tibble(
x1 = 1:1000,
y1 = x1 ** 2
)
plotLinearModel(large_data, 'x1', 'y1')

}
