% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLinearModel.R
\name{plotLinearModel}
\alias{plotLinearModel}
\title{Fit and visualize simple linear model on graph}
\usage{
plotLinearModel(data, x, y)
}
\arguments{
\item{data}{Data set in form of tibble that is being plotted (Must have at least three valid row)}

\item{x}{Column in data that is the x axis of plot (Name of column X should be character (in quotes) and data in column x should be numeric)}

\item{y}{Column in data that is the y axis of plot (Name of column y should be character (in quotes) and data in column y should be numeric)}
}
\value{
ggplot2 scatter plot with linear model and data points with linear
function as the title and adjusted r-squared and F-statistic as subtitle
}
\description{
Plot quantitative data from column x and y in data in a scatter plot along with
the line from the fitted simple linear model. The simple linear function is
the graph title and the subtitle is correlation statistics
(r-squared and F-statistic)
}
\examples{
test_tibble <- tibble::tribble(
~x, ~y, ~z,
#--|--|----
1, 2, 3,
2, 3, 4,
3, 5, 2,
4, 3, 2
)
plotLinearModel(test_tibble, 'x', 'y')

large_data <- tibble::tibble(
x1 = 1:1000,
y1 = x1 ** 2
)
plotLinearModel(large_data, 'x1', 'y1')

}
